+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.4] - 2024-02-23
+### Added
+- Minor style updates
+- Added nprogress to show user a loading state while navigating
+- Add ability to create new account from frontend
+
+### Updated
+- Updated payment to include stripe customer id
+
 ## [0.3.3] - 2024-02-23
 ### Added
 - Sign out to account page